app/requirements.txt
Django
djangorestframework
django-cors-headers
drf-yasg[validation]
psycopg2